'Refactored by Sourcery'

+++ bot/__init__.py
@@ -31,7 +31,7 @@ def get_file_name(message):
 
 @client.on(events.NewMessage)
 async def download(event):
-    if event.is_private :
+    if event.is_private:
         try:
             await event.client(functions.channels.GetParticipantRequest(
                 channel = Config.CHANNEL_USERNAME,
@@ -40,8 +40,8 @@ async def download(event):
         except errors.UserNotParticipantError:
             await event.reply(f"First join to our official channel to access the bot and get the latest updates about the bot\n\n@{Config.CHANNEL_USERNAME}\n\nAfter that send /start agin.")
             return
-        
-        if event.file :
+
+        if event.file:
             sender = await event.get_sender()
             msg = await event.client.send_file(
                 Config.CHANNEL,
@@ -54,7 +54,7 @@ async def download(event):
             return
 
         elif id_msg := re.search("/start (.*)", event.raw_text ):
-            if id_hex := id_msg.group(1) :
+            if id_hex := id_msg.group(1):
                 try:
                     id = int(id_hex,16)
                 except ValueError:
@@ -62,35 +62,36 @@ async def download(event):
                 msg = await event.client.get_messages(Config.CHANNEL,ids=id)
                 if not msg or not msg.file :
                     return await event.reply("404! File Not Found")
-                if regex := re.search(r"(\d*)/(\d*)",msg.message):
-                    if user_id := int(regex.group(1)) :
-                        msg_id = int(regex.group(2))
-                        file = await event.client.get_messages(user_id,ids=msg_id)
-                        if not file or not file.file :
-                            return await event.reply("404! File Not Found")
-                        forward = await file.forward_to(event.chat_id)
-                        id_name = f"{id_hex}/{get_file_name(msg)}"
-                        bot_url = f"https://t.me/{username_bot}?start={id_hex}"
-                        forward_reply = await forward.reply(f"will be deleted in 21 seconds. \n\n📎 : {Config.DOMAIN}/{id_name}\n\n🤖 : {bot_url}",link_preview=False)
-                        await asyncio.sleep(12)
-                        await forward_reply.edit(f"will be deleted in 10 seconds. \n\n📎 : {Config.DOMAIN}/{id_name}\n\n🤖 : {bot_url}")
-                        await asyncio.sleep(10)
-                        await forward.delete()
-                        await forward_reply.edit(f"📎 : {Config.DOMAIN}/{id_name}\n\n🤖 : {bot_url}",link_preview=True)
+                if (regex := re.search(r"(\d*)/(\d*)", msg.message)) and (
+                    user_id := int(regex.group(1))
+                ):
+                    msg_id = int(regex.group(2))
+                    file = await event.client.get_messages(user_id,ids=msg_id)
+                    if not file or not file.file :
+                        return await event.reply("404! File Not Found")
+                    forward = await file.forward_to(event.chat_id)
+                    id_name = f"{id_hex}/{get_file_name(msg)}"
+                    bot_url = f"https://t.me/{username_bot}?start={id_hex}"
+                    forward_reply = await forward.reply(f"will be deleted in 21 seconds. \n\n📎 : {Config.DOMAIN}/{id_name}\n\n🤖 : {bot_url}",link_preview=False)
+                    await asyncio.sleep(12)
+                    await forward_reply.edit(f"will be deleted in 10 seconds. \n\n📎 : {Config.DOMAIN}/{id_name}\n\n🤖 : {bot_url}")
+                    await asyncio.sleep(10)
+                    await forward.delete()
+                    await forward_reply.edit(f"📎 : {Config.DOMAIN}/{id_name}\n\n🤖 : {bot_url}",link_preview=True)
                 return
-        
+
         await event.reply("Send any file to get a link to download it")
-        
+
 
     elif event.is_channel:
-        if event.chat_id == Config.CHANNEL:
-            if event.reply_to:
-                msg = await event.get_reply_message()
-                if regex := re.search(r"(\d*)/(\d*)",msg.message):
-                    if user_id := int(regex.group(1)) :
-                        msg_id = int(regex.group(2))
-                        if await event.client.send_message(entity=user_id, message=event.message, reply_to=msg_id):
-                            await event.client.edit_message(event.chat_id,event.id,f"{event.message.message}\n sended")
+        if event.chat_id == Config.CHANNEL and event.reply_to:
+            msg = await event.get_reply_message()
+            if (regex := re.search(r"(\d*)/(\d*)", msg.message)) and (
+                user_id := int(regex.group(1))
+            ):
+                msg_id = int(regex.group(2))
+                if await event.client.send_message(entity=user_id, message=event.message, reply_to=msg_id):
+                    await event.client.edit_message(event.chat_id,event.id,f"{event.message.message}\n sended")
